+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "src/main/resources/Team_MODI_MODI_BE_KEYSET"]
+	path = src/main/resources/Team_MODI_MODI_BE_KEYSET
+	url = https://github.com/prgrms-web-devcourse/Team_MODI_MODI_BE_KEYSET
