--- keyset.json
@@ -1,1 +1,0 @@
-{"primaryKeyId":825490295,"key":[{"keyData":{"typeUrl":"type.googleapis.com/google.crypto.tink.AesGcmKey","value":"GhA0q6BeSL0OnzEwKn24/jYf","keyMaterialType":"SYMMETRIC"},"status":"ENABLED","keyId":825490295,"outputPrefixType":"TINK"}]}
