+++ src/main/java/com/prgrms/modi/common/configuration/TinkerConfiguration.java
@@ -1,7 +1,9 @@
 package com.prgrms.modi.common.configuration;
 
 import com.google.crypto.tink.aead.AeadConfig;
+import com.google.crypto.tink.integration.awskms.AwsKmsClient;
 import java.security.GeneralSecurityException;
+import java.util.Optional;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
