+++ src/test/java/com/prgrms/modi/party/service/PartyServiceTest.java
@@ -9,7 +9,6 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 import com.prgrms.modi.history.service.CommissionHistoryService;
 import com.prgrms.modi.history.service.PointHistoryService;
@@ -17,9 +16,6 @@ import com.prgrms.modi.ott.domain.OTT;
 import com.prgrms.modi.ott.repository.OttRepository;
 import com.prgrms.modi.party.domain.Party;
 import com.prgrms.modi.party.domain.PartyStatus;
-import com.prgrms.modi.party.dto.request.CreatePartyRequest;
-import com.prgrms.modi.party.dto.request.RuleRequest;
-import com.prgrms.modi.party.dto.response.PartyIdResponse;
 import com.prgrms.modi.party.dto.response.PartyListResponse;
 import com.prgrms.modi.party.repository.PartyRepository;
 import com.prgrms.modi.party.repository.RuleRepository;
@@ -37,11 +33,15 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Pageable;
 
 @ExtendWith(MockitoExtension.class)
 @MockitoSettings(strictness = Strictness.LENIENT)
 class PartyServiceTest {
+
+    private Logger logger = LoggerFactory.getLogger(getClass());
 
     @InjectMocks
     private PartyService partyService;
@@ -111,43 +111,6 @@ class PartyServiceTest {
 
         assertThat(response.getPartyList().size(), equalTo(partySize));
         assertThat(response.getPartyList().get(0).getPeriod(), equalTo(period));
-    }
-
-    @Test
-    @DisplayName("파티를 생성할 수 있다")
-    public void createParty() {
-        // Given
-        Long ottId = 1L;
-        Long userId = 1L;
-        OTT ott = MockCreator.getOttFixture(ottId);
-
-        RuleRequest ruleRequest1 = new RuleRequest(1L, "1인 1회선");
-        RuleRequest ruleRequest2 = new RuleRequest(2L, "양도 금지");
-        List<RuleRequest> ruleRequests = List.of(ruleRequest1, ruleRequest2);
-
-        CreatePartyRequest createPartyRequest = new CreatePartyRequest.Builder()
-            .ottId(ottId)
-            .ottName("넷플릭스")
-            .grade("프리미엄")
-            .partyMemberCapacity(4)
-            .startDate(LocalDate.now())
-            .endDate(LocalDate.now().plusMonths(6))
-            .mustFilled(true)
-            .rules(ruleRequests)
-            .sharedId("modi@gmail.com")
-            .sharedPassword("modimodi123")
-            .build();
-
-        Party party = MockCreator.getPartyFixture(1L);
-
-        when(partyRepository.save(any(Party.class))).thenReturn(party);
-        when(ottRepository.getById(anyLong())).thenReturn(ott);
-
-        // When
-        PartyIdResponse response = partyService.createParty(createPartyRequest, userId);
-
-        // Then
-
     }
 
     @Test
