+++ src/test/resources/application.yaml
@@ -66,3 +66,8 @@ server:
       charset: UTF-8
       enabled: true
       force: true
+
+tink:
+  kms-arn: ${KMS_ARN}
+  aws-kms-access-key-id: ${AWS_KMS_ACCESS_KEY_ID}
+  aws-kms-secret-access-key: ${AWS_KMS_SECRET_ACCESS_KEY}
